requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────
streamlit>=1.32.0

# ── LLM Providers — install the one(s) you plan to use ────────────────────
google-generativeai>=0.5.0   # Google Gemini (gemini-1.5-flash, gemini-1.5-pro …)
openai>=1.12.0               # OpenAI        (gpt-4o, gpt-3.5-turbo …)
groq>=0.9.0                  # Groq          (llama3-70b, mixtral-8x7b …)  [FREE tier]
anthropic>=0.25.0            # Anthropic     (claude-3-5-sonnet, claude-3-opus …)

# ── Bonus: Sentiment Analysis ──────────────────────────────────────────────
textblob>=0.18.0             # Lightweight sentiment — optional but recommended
